requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sentence-transformers
numpy
pypdf
docx2txt
requests
streamlit
